Initial move of raid commands out of main
this version will not run. needs major cleanup
